Push Notification Issue
There is currently an issue with PushNotifications. See Description:

when the user responds to a push notification message, and it is
successfully sent. The device that send the message is also getting
notified, since it is subscribed to push notifications on that channel.

Do you have any advice on a way to address or ignore the user that sent
the message?

Example:

Details:
Channel = “thread1”
Users Subscribed to Receive Push Notifications on Channel = UserA and
UserB

1. UserA Opens the app and publishes a message in the Channel
2. UserB receives an Interactive Push notification since the app on
UserB’s device is not running.
3. UserB responds to the push notification without opening the app.
4. The message is published to the channel using the method:

application(application: UIApplication, handleActionWithIdentifier
identifier: String?, forRemoteNotification userInfo: [NSObject :
AnyObject], withResponseInfo responseInfo: [NSObject : AnyObject],
completionHandler: () -> Void)

5. UserA receives a Push Notification or a message in the channel
depending on the state of the app on their device.

The Problem:

UserB Receives a push notification about the message they just
published, since they never opened their app.

Potential Solution:

Each user subscribes to their own channel, the thread channel is
actually a channel group. The channels for the user in a conversation
will be added to  the thread channel. The conversation messages will be
published to the group. However push notifications will be published to
specified users.

Details:
UserA channel = UserAChannel
UserB channel = UserBChannel

Thread = UserAChannel | UserBChannel

Setup

	1. UserA has subscribed to UserAChannel
	2. UserB is has subscribed to UserBChannel
	3. UserA has added Publish to UserBChannel
	4. UserB has added Publish to UserAChannel
	5. UserA has added Push Notification to UserAChannel
	6. UserB has added Push Notification to UserBChannel
	7. UserA and UserB subscribe and publish to Thread

Message

	1. Loading Messages In View:
		a. Only Thread messages are loaded in view.
	2. Sending:
		a. UserA publishes a message to Thread - No Push payload
		b. UserA publishes a message to UserBChannel Only PushPayload
	3. Receiving:
		a. UserB receives the push notification if app is not active since
UserB is setup to receive Push Notifications on the channel UserBChannel
		b. UserB receives the message in the app if the app is open.
(Messages loaded into view only from Thread)
	4. Responding To Push (Interactive):
		a. UserB replies to push notification and publishes a message to
Thread - No Push payload
		b. UserB reaplies to push notification and publishes a message to
UserAChannel Only PushPayload

Concern:

	1. In integration, would need to work out logic regarding when to
setup 3 and 4 Setup steps listed above. It would have to be clear as to
when the opposite user channel should be exposed to the user requesting
it.