Fourth Update:
Fourth Update:

Summary:

LOAD Message was already updated as part of a previous story, app is in
working condition (TESTED). Next Step work on Push Message.

	1. Update Load Message to pull, parse and display the new message
packet that does not include the APN and GCM payloads:
		a. Loading Message
			i. Load Messages from Thread - Update this to only message Packet,
without APN or GCM
			ii. didRecieveMessage will have to be updated to parse the new
message packet that will not have APN or GCM

Summary:

LOAD Message was already updated as part of a previous story, app is in
working condition (TESTED). Next Step work on Push Message.

On - Going - Concern:

	1. In integration, would need to work out logic regarding when to
setup 3 and 4 Setup steps listed above. It would have to be clear as to
when the opposite user channel should be exposed to the user requesting
it.
Currently in didPressRightButton hard coded logic to send push to
specific channel.