Fifth Update:
Fifth Update:

Summary:

Updated Push Message, app is in working condition (TESTED).

	1. Update Push Notification handling:
		a. Interaction and Action Handling
			i. Update handleActionWithIdentifier
			ii. Update handleMeesageResponse
			iii. Update sendMessage
			iv. create sendPushMessage
		b. Update handleActionWithIdentifier
			i. Subscribe the user to current User's channel and to the Thread
		c. Update handleMeesageResponse
			i. Change the current sendMessage(channel : String, messagePacket :
[String : String]) method to sendPushMessage
			ii. Create a new sendMessage method that sends payload without APNs
GCM data
			iii. Call sendPushMessage and pass in the channel name of the
intended user
			iv. Call sendMessage and pass in the channel name of the thread
		d. Update sendMessage
			i. Create a new method called sendPushMessage and move the code from
the current send message function in there.
			ii. Change the method signature to ask for threadName instead of
just channel
			iii. change the message packet to send only message payload without
APN and GCM data
		e. Create sendPushMessage
			i. Change the method signature to ask for IntendedUsersChannelName
instead of just channel
			ii. Move the code from the previous sendMessage method
Summary:

Updated Push Message, app is in working condition (TESTED).

New - Concern:

	1. In integration, would need to work out logic regarding when to
setup 3 and 4 Setup steps listed above. It would have to be clear as to
when the opposite user channel should be exposed to the user requesting
it.
		a. Currently in didPressRightButton hard coded logic to send push to
specific channel.
Sometimes the Interactive Push Notification response is not working,
specifically when the device is connected to the computer and app is
running from XCODE.